Correcoes P0/P1 e melhorias estruturais (P2/docs)
Bugfixes:
- Mod 07: InitializingBean para criar tabela automaticamente
- Mod 04: JSP legado removido, links corrigidos
- Mod 05: rollback com restauracao de backup de saldo
- Mod 10: status PENDENTE em dados iniciais
- Mod 11: @Async refatorado (evita recursao no this), condicao de corrida do retry corrigida

Melhorias:
- Mod 02/03: e.printStackTrace() substituido por java.util.logging
- Mod 07/09/10/11/12: double -> BigDecimal para valores monetarios
- Mod 09/10/11: DTOs para desacoplar API do modelo interno
- Mod 09/10/11/12: testes unitarios com JUnit 5 + Mockito
- Mod 12: pom.xml corrigido para Java 21 (pattern matching final)
- data-model.md: tipos atualizados (BigDecimal, records do Mod 12)
- backlog.md: registro de todos os itens concluidos

+++ 07-spring-xml-service-layer/src/pedidos/Pedido.java
@@ -1,16 +1,18 @@
 package pedidos;
 
+import java.math.BigDecimal;
+
 public class Pedido {
     private int id;
     private String cliente;
     private String produto;
     private int quantidade;
-    private double valorTotal;
+    private BigDecimal valorTotal;
     private String status; // "PENDENTE", "PAGO", "CANCELADO"
 
     public Pedido() {}
 
-    public Pedido(String cliente, String produto, int quantidade, double valorTotal) {
+    public Pedido(String cliente, String produto, int quantidade, BigDecimal valorTotal) {
         this.cliente = cliente;
         this.produto = produto;
         this.quantidade = quantidade;
@@ -18,7 +20,6 @@ public class Pedido {
         this.status = "PENDENTE";
     }
 
-    // getters and setters for all fields
     public int getId() { return id; }
     public void setId(int id) { this.id = id; }
     public String getCliente() { return cliente; }
@@ -27,8 +28,8 @@ public class Pedido {
     public void setProduto(String produto) { this.produto = produto; }
     public int getQuantidade() { return quantidade; }
     public void setQuantidade(int quantidade) { this.quantidade = quantidade; }
-    public double getValorTotal() { return valorTotal; }
-    public void setValorTotal(double valorTotal) { this.valorTotal = valorTotal; }
+    public BigDecimal getValorTotal() { return valorTotal; }
+    public void setValorTotal(BigDecimal valorTotal) { this.valorTotal = valorTotal; }
     public String getStatus() { return status; }
     public void setStatus(String status) { this.status = status; }
 }
