+++ 12-modern-java-language-evolution/src/main/java/regras/Solicitacao.java
@@ -1,11 +1,13 @@
 package regras;
 
+import java.math.BigDecimal;
+
 public sealed interface Solicitacao permits Solicitacao.Emprestimo, Solicitacao.Credito, Solicitacao.Consorcio {
 
-    record Emprestimo(String cliente, double valor, int parcelas, double rendaMensal) implements Solicitacao {}
-    record Credito(String cliente, double valor, double limiteDisponivel, boolean possuiRestricao) implements Solicitacao {}
-    record Consorcio(String cliente, double valor, int totalParcelas, int parcelasPagas) implements Solicitacao {}
+    record Emprestimo(String cliente, BigDecimal valor, int parcelas, BigDecimal rendaMensal) implements Solicitacao {}
+    record Credito(String cliente, BigDecimal valor, BigDecimal limiteDisponivel, boolean possuiRestricao) implements Solicitacao {}
+    record Consorcio(String cliente, BigDecimal valor, int totalParcelas, int parcelasPagas) implements Solicitacao {}
 
     String cliente();
-    double valor();
+    BigDecimal valor();
 }
